Shorten the names of some long name headers
