+++ sdk/keyvault/azure-security-keyvault-administration/src/settings_client.cpp
@@ -11,7 +11,7 @@
 #include <azure/core/internal/json/json_optional.hpp>
 #include <azure/core/internal/json/json_serializable.hpp>
 #include <azure/keyvault/administration/settings_client.hpp>
-#include <azure/keyvault/shared/keyvault_challenge_based_authentication_policy.hpp>
+#include <azure/keyvault/shared/keyvault_challenge_based_auth.hpp>
 #include <azure/keyvault/shared/keyvault_shared.hpp>
 
 #include <memory>
