+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/cryptography_client.cpp
@@ -17,7 +17,7 @@
 #include <azure/core/exception.hpp>
 #include <azure/core/http/http.hpp>
 #include <azure/core/http/policies/policy.hpp>
-#include <azure/keyvault/shared/keyvault_challenge_based_authentication_policy.hpp>
+#include <azure/keyvault/shared/keyvault_challenge_based_auth.hpp>
 #include <azure/keyvault/shared/keyvault_shared.hpp>
 
 #include <algorithm>
