+++ sdk/keyvault/azure-security-keyvault-keys/src/key_client.cpp
@@ -15,7 +15,7 @@
 #include <azure/core/http/http.hpp>
 #include <azure/core/http/policies/policy.hpp>
 #include <azure/core/internal/http/pipeline.hpp>
-#include <azure/keyvault/shared/keyvault_challenge_based_authentication_policy.hpp>
+#include <azure/keyvault/shared/keyvault_challenge_based_auth.hpp>
 #include <azure/keyvault/shared/keyvault_shared.hpp>
 
 #include <algorithm>
