+++ sdk/keyvault/azure-security-keyvault-secrets/src/secret_client.cpp
@@ -18,7 +18,7 @@
 #include <azure/core/credentials/credentials.hpp>
 #include <azure/core/http/http.hpp>
 #include <azure/core/http/policies/policy.hpp>
-#include <azure/keyvault/shared/keyvault_challenge_based_authentication_policy.hpp>
+#include <azure/keyvault/shared/keyvault_challenge_based_auth.hpp>
 #include <azure/keyvault/shared/keyvault_shared.hpp>
 
 #include <algorithm>
